Use pickle instead of deepcopy to copy regex objects
The `regex` package doesn't actually implement `__deepcopy__`. It
really implements a fake and returns just a reference to the same
`regex` object. In order to make actual copies, the pickle package
is used.
To facilitate making copies, the compiled regex is stored in the
`Rule` as `self.pickled_regex`, so a new instance of it can be
directly created with `pickle.loads()`.